+++ src/main.rs
@@ -2,12 +2,12 @@ use socarel::{Forest, Tree, Node, NodeContent, TreeIdentifier};
 
 fn main() {
     let mut forest = <Forest>::new();
-    forest.new_tree("my_tree");
-    forest.new_tree("other_tree");
+    forest.new_tree("my_tree").unwrap();
+    forest.new_tree("other_tree").unwrap();
 
     println!("Forest = {:#?}", forest);
 
-    if let Some(my_tree) = forest.get_mut_tree("my_tree") {
+    if let Ok(my_tree) = forest.get_mut_tree("my_tree") {
         let _root = my_tree.set_root("my root node").unwrap();
         let _child_1 = my_tree.link_node("bad child node 1", _root).unwrap();
         let _child_2 = my_tree.link_node("child node 2", _root).unwrap();
@@ -22,7 +22,7 @@ fn main() {
 
         println!("-------------------------------------------------------");
 
-        my_tree.update_node("child node 1", _child_1);
+        my_tree.update_node("child node 1", _child_1).unwrap();
 
         println!("New Child 1 content = {:#?}", my_tree.get_node_content(_child_1).unwrap().get_val());
 
@@ -30,7 +30,7 @@ fn main() {
 
         println!("-------------------------------------------------------");
 
-        my_tree.unlink_node(_child_1);
+        my_tree.unlink_node(_child_1).unwrap();
 
         println!("My Tree after unlink = {:#?}", my_tree);
 
@@ -123,6 +123,8 @@ fn main() {
 
     println!("--- In-Order DFS Iter:");
     iterate(tree.iterators().in_dfs());
+    println!("--- Inverse In-Order DFS Iter:");
+    iterate(tree.iterators().inv_in_dfs());
 }
 
 fn iterate<'a>(iter: impl Iterator<Item=(&'a Node, usize)>) {