+++ CHANGELOG.md
@@ -11,10 +11,12 @@ All notable changes to the **Socarel** project will be documented in this file.
 - New function to find paths, `Tree::find_path`.
 - Trait for Tree IDs, `TreeIdentifier`, similar to NodeContent. Also added a macro (`impl_tree_id_traits!`) to implement default details.
 - In-order DFS iterator.
+- Error type.
 
 ### Changes
 
 - Deprecated `Tree::find_node`.
+- Some functions that returned Option now return Result.
 
 ### Impoved